+++ vector/src/main/java/im/vector/riotx/core/platform/VectorBaseFragment.kt
@@ -234,9 +234,8 @@ abstract class VectorBaseFragment : BaseMvRxFragment(), HasScreenInjector {
 
     private val uiDisposables = CompositeDisposable()
 
-    protected fun Disposable.disposeOnDestroyView(): Disposable {
+    protected fun Disposable.disposeOnDestroyView() {
         uiDisposables.add(this)
-        return this
     }
 
     /* ==========================================================================================